+++ inbm/Changelog.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - MMMM-DD-YY
-
+### Changed
+ - (NEXMANAGE-874) Remove UT rollback command in dispatcher agent
 
 ## 4.2.6 - 2024-10-04
 ### Added